Add boot page. Font not entirely working.
