+++ app/src/app/dinner-list/dinner-list.component.scss
@@ -1,7 +1,3 @@
-table {
-    width: 100%;
-}
-
 .vertical-align-middle {
     vertical-align: middle;
 }
