+++ app/src/app/services/dinner-list.service.ts
@@ -1,6 +1,7 @@
 import {Injectable} from '@angular/core';
 import {HttpClient} from '@angular/common/http';
 import {environment} from '../../environments/environment';
+import {IDinnerDate} from './models/dinner.models';
 
 @Injectable({
     providedIn: 'root'
@@ -9,26 +10,16 @@ export class DinnerListService {
     API_URL = environment.baseUrl;
     URL_WEEK = `${this.API_URL}/dinnerweek/`;
 
+    dayNames = ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday'];
+
     constructor(private  httpClient: HttpClient) {
         // console.log(this.URL_BASE);
-        this.getDinnerWeek().then(result => {
-            console.log(result);
-        });
     }
 
     getDinnerWeek() {
         return this.httpClient.get<IDinnerDate[]>(`${this.URL_WEEK}`).toPromise();
     }
 
-}
 
-export interface IDinnerDate {
-    id: number;
-    num_eating: number;
-    cost: number;
-    open: boolean;
-    date: Date;
 
-    signup_time: Date;
-    close_time: Date;
 }
