+++ app/src/app/services/models/dinner.models.ts
@@ -1,0 +1,12 @@
+export interface IDinnerDate {
+    id: number;
+    num_eating: number;
+    cost: number;
+    open: boolean;
+    date: string;
+
+    cook: string;
+
+    signup_time: Date;
+    close_time: Date;
+}
