+++ app/src/index.html
@@ -13,7 +13,26 @@
     <meta name="theme-color" content="#1976d2">
 </head>
 <body>
-<app-root></app-root>
+<app-root>
+    <!-- Angular fallback -->
+    <link rel="stylesheet" href="assets/css/angular-fallback.min.css">
+    <h2 class="display-4">Loading DS4 app</h2>
+    <!--    <h2> Who needs an app anyway</h2>-->
+    <!--    <br/>-->
+
+    <div class="app-loading">
+        <div class="logo-fallback"></div>
+        <svg class="spinner" viewBox="25 25 50 50">
+            <circle class="path" cx="50" cy="50" r="15" fill="none" stroke-width="2" stroke-miterlimit="20"/>
+        </svg>
+    </div>
+
+    <div class="fab">
+        <span class="fab-action-button" value="Reload page" onClick="window.location.reload()">
+            <i class="fab-action-button__icon"></i>
+        </span>
+    </div>
+</app-root>
 <noscript>Please enable JavaScript to continue using this application.</noscript>
 </body>
 </html>
