+++ app/src/manifest.webmanifest
@@ -2,7 +2,7 @@
   "name": "app",
   "short_name": "app",
   "theme_color": "#03A9F4",
-  "background_color": "#fafafa",
+  "background_color": "#03A9F4",
   "display": "standalone",
   "scope": "/",
   "start_url": "/",
